refactor: Moved state layer to new component
